# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "Lorem Ipsum",
        "contact": {
            "email": "hi.wasissubekti02@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/banner": {
            "get": {
                "tags": [
                    "Banner"
                ],
                "summary": "ambil semua banner",
                "description": "Mengambil Data Semua banner",
                "operationId": "banner",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "success get all banner",
                                    "data": [
                                        {
                                            "id": 15,
                                            "image": "1801733198090700.jpg",
                                            "description": "bio",
                                            "created_at": "2024-06-10T02:36:13.000000Z",
                                            "updated_at": "2024-06-11T08:01:32.000000Z"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Banner"
                ],
                "summary": "Create new banner",
                "description": "Create a new banner",
                "operationId": "create_banner",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "The image of banner",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "description": {
                                        "description": "The description of banner",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": true,
                                    "message": "success post banner",
                                    "data": {
                                        "image": "1801799915486786.jpg",
                                        "description": "lorem ipsum",
                                        "created_at": "2024-06-14T01:21:56.000000Z",
                                        "updated_at": "2024-06-14T01:21:56.000000Z",
                                        "id": 21
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": false,
                                    "message": "fail post banner",
                                    "error": "Validation error message"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/banner/{id}": {
            "get": {
                "tags": [
                    "Banner"
                ],
                "summary": "Get single banner",
                "description": "Retrieve a single banner by ID",
                "operationId": "single_banner",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the banner",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": true,
                                    "message": "success get blog",
                                    "data": {
                                        "id": 15,
                                        "image": "1801556148382093.jpg",
                                        "description": "ok",
                                        "created_at": "2024-06-10T02:36:13.000000Z",
                                        "updated_at": "2024-06-13T08:17:59.000000Z"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "banner not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": false,
                                    "message": "fail get banner",
                                    "error": "No query results for model [App\\\\Models\\\\Banner] 10"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Banner"
                ],
                "summary": "Update banner",
                "description": "Update banner by ID",
                "operationId": "update_banner",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of banner",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "description": "The description of banner",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "The image of banner",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": true,
                                    "message": "success update banner",
                                    "data": {
                                        "id": 1,
                                        "description": "Updated banner",
                                        "image": "updated_image_name.jpg",
                                        "created_at": "2024-05-18 15:52:01",
                                        "updated_at": "2024-05-18 15:52:01"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "banner not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": false,
                                    "message": "fail update banner",
                                    "error": "No query results for model [App\\\\Models\\\\Banner] 10"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Banner"
                ],
                "summary": "Delete banner",
                "description": "Delete a blog banner by ID",
                "operationId": "delete_banner",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the banner",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": true,
                                    "message": "success delete banner"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "banner not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": false,
                                    "message": "fail delete banner",
                                    "error": "No query results for model [App\\\\Models\\\\Banner] 10"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}